+++ app.js
@@ -20,6 +20,7 @@ const prom = require('./utils/prometheus');
 const swaggerSpec = require('./swagger.json');
 const swaggerUi = require('swagger-ui-express');
 const signoffController = require('./components/signoffs/signoffController');
+const authenticationController = require('./utils/apiAuth');
 
 const app = express();
 const port = process.env.PORT || 3000;
@@ -112,6 +113,8 @@ app.get('/metrics', async (req, res) => {
 app.post(`/case/:caseIdentifier/${urls.signoff}`, signoffController.addSignoff);
 app.post(`/case/${urls.signoff}`, signoffController.addBatchSignoffs);
 app.get(`/case/:caseIdentifier/${urls.signoff}`, signoffController.getSignoff);
+
+app.post('/token', authenticationController.addNewKey);
 
 app.use(errorHandler);
 app.use(prom.monitorAfterRequest);